Add a comment explaining the CLike impls
Generally, to_u32 and from_u32 are a cast and a transmute,
respectively, but here we need to manually "encode" the values since
the enum values are mapped to the relevant bits for libpcre.

If everyone's alright with it, the bitflags! macro seems to be a
more elegant way to handle this problem.